Rename profile routes
