+++ src/components/App.jsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from 'react';
-import { Route, Routes, useLocation, useNavigate  } from 'react-router-dom';
+import { Navigate, Route, Routes, useLocation, useNavigate  } from 'react-router-dom';
 import { PrivacyPolicy } from './PrivacyPolicy';
 import { MyProfile } from './MyProfile';
-import { MyProfileNew } from './MyProfileNew';
+import { MyProfileOld } from './MyProfileOld';
 import { LoginScreen } from './LoginScreen';
 import { SubmitForm } from './SubmitForm';
 import { AddNewProfile } from './AddNewProfile';
@@ -87,8 +87,9 @@ export const App = () => {
       <Route path="/" element={canAccessAdd ? <AddNewProfile isLoggedIn={isLoggedIn} setIsLoggedIn={setIsLoggedIn} /> : <PrivacyPolicy />} />
       <Route path="/login" element={<LoginScreen isLoggedIn={isLoggedIn} setIsLoggedIn={setIsLoggedIn} />} />
       <Route path="/submit" element={<SubmitForm />} />
-      <Route path="/my-profile"  element={<MyProfile isLoggedIn={isLoggedIn} setIsLoggedIn={setIsLoggedIn}/>} />
-      <Route path="/my-profile-new" element={<MyProfileNew />} />
+      <Route path="/my-profile" element={<MyProfile />} />
+      <Route path="/my-profile-new" element={<Navigate to="/my-profile" replace />} />
+      {isAdmin && <Route path="/my-profile-old" element={<MyProfileOld isLoggedIn={isLoggedIn} setIsLoggedIn={setIsLoggedIn} />} />}
       {canAccessAdd && <Route path="/add" element={<AddNewProfile isLoggedIn={isLoggedIn} setIsLoggedIn={setIsLoggedIn} />} />}
       {canAccessMatching && <Route path="/matching" element={<Matching />} />}
       {isAdmin && <Route path="/edit/:userId" element={<EditProfile />} />}
